Rename test-data to test-vectors
